NIFI-2292: Funnel all cluster node status changes through the cluster coordinator instead of having each node broadcast changes to the whole cluster. This gives us the ability to increment the updateId consistently without race conditions.

+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-api/src/main/java/org/apache/nifi/web/api/ControllerResource.java
@@ -548,6 +548,10 @@ public class ControllerResource extends ApplicationResource {
                     + "not equal the node id of the requested resource (%s).", requestNodeDTO.getNodeId(), id));
         }
 
+        if (isReplicateRequest()) {
+            return replicateToCoordinator(HttpMethod.PUT, nodeEntity);
+        }
+
         // update the node
         final NodeDTO node = serviceFacade.updateNode(requestNodeDTO);
 
@@ -598,6 +602,10 @@ public class ControllerResource extends ApplicationResource {
         // ensure connected to the cluster
         if (!isConnectedToCluster()) {
             throw new IllegalClusterResourceRequestException("Only a node connected to a cluster can process the request.");
+        }
+
+        if (isReplicateRequest()) {
+            return replicateToCoordinator(HttpMethod.DELETE, getRequestParameters());
         }
 
         serviceFacade.deleteNode(id);
